requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Pillow==10.2.0
supabase==2.3.4
python-jose==3.3.0
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
django-filter==23.5
drf-yasg==1.21.7
black==24.1.1
pytest==8.0.0
pytest-django==4.8.0
factory-boy==3.3.0
coverage==7.4.1
dj-database-url==2.1.0
requests==2.31.0
PyJWT==2.8.0 